+++ index.js
@@ -11,6 +11,7 @@ export * as vec2 from "./vec2.js";
 export * as vec3 from "./vec3.js";
 export * as vec4 from "./vec4.js";
 
+export * as avec2 from "./avec2.js";
 export * as avec3 from "./avec3.js";
 export * as avec4 from "./avec4.js";